segregate animated and physics bodies into two separate lists
we construct a vector of pointers to all of them ( careful w cache misses, should force the two vectors to be allocated consecutively ), and pass THAT vector to the renderer so it renders both animated bodies and physics bodies indescriminantly.

this exempts the animated bodies from having collision, though they have all teh same data so they could be manually added to teh physics system later on if desired.